Force ajax dataType to html
This fixes the "html.replace is not a function" error.

`utility.storePageIn` requires the 3rd parameter to be a jQuery or a string. But, on line ~452, the ajax success callback caller passes an XMLDocument (instead off an HTML string) object to the `html` parameter in the success (in my case, at least). This is passed to `ulitity.storePageIn` which expects a jQuery or a string.

This change forces the request result to be an HTML string, instead of the default auto-guessing.