Fix back-button not loading hash urls
Fixes #57 - pop state does not load the previous page if that page has a hash in the url.

Changed `onPopState` and `isHash`.

`isHash` gets the source url from `window.location.pathname` but it no longer holds the source url when `onpopstate` is triggered. `window.location.pathname` already holds the destination url on `onpopstate`.

I made it so that `onPopState` expilicitly tells `isHash` what the source url is to compare with the destination url.