New option: hrefRegex
hrefRegex is a regular expression to specify which anchors with a
specific href property (based on the regex) smoothState should bind to.
If empty, every href will be permitted.

For example. In my application, I only need to bind smoothState on the
links that have "f?p=" in the URL. So, with this, I can do it.